+++ src/hooks/useFetchLocations.ts
@@ -1,12 +1,13 @@
 import { useMap } from "react-leaflet";
 import { ChannelData } from "@/types";
 import { useRouter } from "next/router";
-import { Dispatch, SetStateAction } from "react";
+import { Dispatch, SetStateAction, useEffect } from "react";
 import { isValidReasons } from "@/utils/isValidReasons";
 import { getFetchAreaBounds } from "@/utils/getFetchAreaBounds";
 import { useSingletonsStore } from "@/stores/singletonsStore";
 import useSWR from "swr";
 import { useDebounce } from "use-debounce";
+import { useLoading } from "@/stores/loadingStore";
 
 const REQUEST_DEBOUNCE_TIME = 500; //ms
 
@@ -18,10 +19,11 @@ const useFetchLocations = (
   const queryReasons = router.query.reasons as string | undefined;
   const bounds = map.getBounds();
   const { apiClient } = useSingletonsStore();
+  const { showLoading } = useLoading();
 
   const debouncedParams = useDebounce([queryReasons], REQUEST_DEBOUNCE_TIME);
 
-  useSWR(
+  const { isLoading } = useSWR(
     // sends request when some item has changed at the below
     ["areas", ...debouncedParams],
     () => {
@@ -40,6 +42,10 @@ const useFetchLocations = (
       },
     }
   );
+
+  useEffect(() => {
+    showLoading(isLoading);
+  }, [isLoading, showLoading]);
 };
 
 export { useFetchLocations };
