+++ src/newlayout/components/CooldownButtonComponent/CooldownButtonComponent.tsx
@@ -1,28 +1,12 @@
+import { useLoading } from "@/stores/loadingStore";
 import { Button, LinearProgress, SxProps, Theme } from "@mui/material";
 import { useTranslation } from "next-i18next";
 import { mutate } from "swr";
-import { create } from "zustand";
-import { devtools } from "zustand/middleware";
 
 interface IStyles {
   [key: string]: SxProps<Theme>;
 }
 
-interface IUseLoading {
-  loading: boolean;
-  showLoading: (_state: boolean) => void;
-}
-
-export const useLoading = create<IUseLoading>()(
-  devtools(
-    (set) => ({
-      loading: false,
-      showLoading: (state: boolean) =>
-        set(() => ({ loading: state }), undefined, { type: "showLoading" }),
-    }),
-    { name: "LoadingStore" }
-  )
-);
 export const CooldownButtonComponent = () => {
   const { t } = useTranslation("home");
   const { loading } = useLoading();
